Move tests into the actual module
