+++ crates/module-system/module-implementations/integration-tests/Cargo.toml
@@ -49,7 +49,6 @@ sov-attester-incentives = { workspace = true, features = ["native"] }
 sov-chain-state = { workspace = true, features = ["native"] }
 sov-value-setter = { workspace = true, features = ["native"] }
 sov-test-modules = { workspace = true, features = ["native"] }
-sov-timelock = { workspace = true, features = ["native"] }
 sov-sequencer-registry = { workspace = true, features = ["native"] }
 sov-bank = { workspace = true, features = ["native"] }
 sov-accounts = { workspace = true, features = ["native"] }
