+++ crates/module-system/module-implementations/sov-timelock/tests/integration/main.rs
@@ -1,0 +1,3 @@
+mod timelock;
+
+type S = sov_test_utils::TestSpec;
